fix: onboarding page protection
